mini_filebot_msi_ci_template/MiniFileBot/requirements.txt
PySide6
requests
PyYAML
guessit
tqdm
python-dateutil
